third_party/WebKit/LayoutTests/fast/dom/shadow/all-in-shadow-tree.html

<!DOCTYPE html>
<script src="../../../resources/js-test.js"></script>
<script src="resources/shadow-dom.js"></script>
<style>
.test { width: 100px; height: 100px; }
</style>
<div class="test"></div>
<div class="test"></div>
<script>
function insertIntoShadowRoot(element) {
  var shadowRoot = element.createShadowRoot();
  shadowRoot.innerHTML = '<div style="all: initial; display: block; width:100%; height: 100%; background: black"></div>';
}
var tests = document.querySelectorAll('.test');
insertIntoShadowRoot(tests[0]);
insertIntoShadowRoot(tests[1]);

description('Test for issue 432257: Matched properties cache should work for all property.');
var targets = document.querySelectorAll('.test::shadow > div');
shouldBe('window.getComputedStyle(targets[0]).width', '"100px"');
shouldBe('window.getComputedStyle(targets[1]).width', '"100px"');
</script>
