third_party/WebKit/LayoutTests/fast/dom/shadow/all-in-shadow-tree-expected.txt

Test for issue 432257: Matched properties cache should work for all property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.getComputedStyle(targets[0]).width is "100px"
PASS window.getComputedStyle(targets[1]).width is "100px"
PASS successfullyParsed is true

TEST COMPLETE
